Fixes refraction

+++ src/color/color_op_2.c
@@ -6,7 +6,7 @@
 /*   By: chaueur <chaueur@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/08 17:42:47 by chaueur           #+#    #+#             */
-/*   Updated: 2017/12/05 12:22:10 by chaueur          ###   ########.fr       */
+/*   Updated: 2017/12/08 08:26:20 by chaueur          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void				color_div_fac(t_color *c, int fac)
 	}
 }
 
-void				color_mult_fac(t_color *c, int fac)
+void				color_mult_fac(t_color *c, double fac)
 {
 	if (c)
 	{
